+++ src/main/java/com/zazaero/data/dto/indCfg/PutMemInfoDTO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class PutMemInfoDTO {
 
-    private Long ind_cfg_uid;
+    private Integer ind_cfg_uid;
     private String cfg_val1;
 }
