+++ src/main/java/com/zazaero/data/dto/indCfg/PutPayBankDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class PutPayBankDTO {
 
-    private Long ind_cfg_uid;
+    private Integer ind_cfg_uid;
     private String cfg_val1;
     private String cfg_val2;
     private String cfg_val3;
